Reduce "extra closing token" math. Use methods. Fixes.
Uses BlackLabIndex.{en|de}codeTokenLengthField() instead of
adding/subtracting manually. Tries to decode the value as soon as
possible so we're mostly working with the actual document length.

A few places didn't handle the extra token correctly, now fixed.

+++ engine/src/main/java/nl/inl/blacklab/forwardindex/AnnotationForwardIndexGlobal.java
@@ -100,7 +100,7 @@ public class AnnotationForwardIndexGlobal implements AnnotationForwardIndex {
     public long docLength(int docId) {
         LeafReaderContext lrc = leafReaderLookup.forId(docId);
         AnnotationForwardIndex fi = FieldForwardIndex.get(lrc, luceneField);
-        return (int)fi.docLength(docId - lrc.docBase);
+        return fi.docLength(docId - lrc.docBase);
     }
 
     @Override
